Move KeyMap to it's own module and make keyBindingManager it's own module
